requirements.txt
# Core AI and Database Requirements
openai==1.107.1
python-dotenv==1.1.1

# Data Processing
pandas==2.3.2
numpy==2.3.3

# Web Dashboard
streamlit==1.49.1
plotly==6.3.0

# Additional Dependencies (automatically installed with above packages)
# requests>=2.32.0
# pydantic>=2.11.0
# typing-extensions>=4.15.0